pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "hl-maker-bot"
version = "0.1.0"
description = "Hyperliquid maker-centric market making bot with local web dashboard"
requires-python = ">=3.11"
dependencies = [
    "hyperliquid-python-sdk>=0.22.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "jinja2>=3.1.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "aiofiles>=23.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",   # for FastAPI test client
    "pytest-cov>=5.0.0",
]

[project.scripts]
hl-bot = "app.main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests that require network access (deselect with -m 'not integration')",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
